Update to use new method of passing password
With later versions of libunrar the method used to supply the password to an archive through the callback changed. Previously you called RARSetPassword, however that does not work with header passwords, so they came up with a different approach. The library passes the address of a wchar character buffer and its size as parameterOne and parameterTwo, and you need to copy the password string into that buffer.